- update test service * postService

+++ src/main/java/rest/api/rest_service/service/PostService.java
@@ -2,10 +2,12 @@ package rest.api.rest_service.service;
 
 import rest.api.rest_service.dao.impl.PostDaoImpl;
 import rest.api.rest_service.entity.PostEntity;
+import rest.api.rest_service.exception.DaoException;
 import rest.api.rest_service.service.dto.PostDtoIn;
 import rest.api.rest_service.service.dto.PostDtoOut;
 
 import java.util.List;
+import java.util.NoSuchElementException;
 
 public class PostService {
     private final static PostService INSTANCE = new PostService();
@@ -28,11 +30,16 @@ public class PostService {
     }
 
     public PostDtoOut findById(Long id) {
-        return postDao.findById(id).stream()
-                .map(entry -> new PostDtoOut(entry.getId(),
-                        entry.getTitle()))
-                .findAny()
-                .orElseThrow();
+        try {
+            return postDao.findById(id).stream()
+                    .map(entry -> new PostDtoOut(entry.getId(),
+                            entry.getTitle()))
+                    .findAny()
+                    .orElseThrow();
+        } catch (NoSuchElementException e) {
+            throw new DaoException("Найти не удалось проверьте верны ли параметры");
+        }
+
     }
 
     public PostDtoOut save(PostDtoIn postDtoIn) {
@@ -43,8 +50,12 @@ public class PostService {
                 id, postDtoIn.getTitle());
     }
 
-    public void deleteById(Long id) {
-        postDao.deleteById(id);
+    public boolean deleteById(Long id) {
+        if(postDao.findById(id).isPresent()) {
+            postDao.deleteById(id);
+            return true;
+        }
+        return false;
     }
 
     public void deleteAll() {
@@ -53,11 +64,16 @@ public class PostService {
         }
     }
 
-    public void update(PostDtoIn postDtoIn) {
-        postDao.update(new PostEntity(
-                postDtoIn.getId(),
-                postDtoIn.getTitle()
-        ));
+    public boolean update(PostDtoIn postDtoIn) {
+        if (postDtoIn != null) {
+            postDao.update(new PostEntity(
+                    postDtoIn.getId(),
+                    postDtoIn.getTitle()
+            ));
+            return true;
+        }
+
+        return false;
     }
